fix little bug from table comment

+++ src/main/java/com/cys/community/dto/QuestionDTO.java
@@ -13,9 +13,9 @@ public class QuestionDTO {
     private Long id;
     private String title;
     private String description;
-    private long gmtCreate;
-    private long gmtModified;
-    private Integer creator;
+    private Long gmtCreate;
+    private Long gmtModified;
+    private Long creator;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
